维护工具

+++ bio-drqi-bsm/bio-drqi-bsm-dto/src/main/java/com/bio/drqi/bsm/rsp/BmsProductStockDetailRspDTO.java
@@ -57,17 +57,17 @@ public class BmsProductStockDetailRspDTO {
     /**
      * 累计入库数量
      */
-    private Integer totalStoreNumber;
+    private BigDecimal totalStoreNumber;
 
     /**
      * 当前库存数量
      */
-    private Integer currentStockNumber;
+    private BigDecimal currentStockNumber;
 
     /**
      * 累计出库数量
      */
-    private Integer totalOutNumber;
+    private BigDecimal totalOutNumber;
 
 
 
@@ -93,7 +93,7 @@ public class BmsProductStockDetailRspDTO {
 
     private String expirationDate;
 
-    private Integer returnNumber;
+    private BigDecimal returnNumber;
 
     private String stockCode;
