+++ bio-drqi-bsm/bio-drqi-bsm-dto/src/main/java/com/bio/drqi/bsm/rsp/BmsProductStockOutLogDetailRspDTO.java
@@ -59,7 +59,7 @@ public class BmsProductStockOutLogDetailRspDTO {
     /**
      * 出库数量
      */
-    private Integer outNumber;
+    private BigDecimal outNumber;
 
     /**
      * 申请人ID
