+++ bio-drqi-bsm/bio-drqi-bsm-dto/src/main/java/com/bio/drqi/bsm/rsp/BmsProductStockOutLogListPageRspDTO.java
@@ -62,7 +62,7 @@ public class BmsProductStockOutLogListPageRspDTO {
     /**
      * 出库数量
      */
-    private Integer outNumber;
+    private BigDecimal outNumber;
 
     /**
      * 申请人ID
